requirements.txt
asgiref==3.8.1
coverage==7.6.4
dj-database-url==2.3.0
Django==5.1.3
djangorestframework==3.15.2
exceptiongroup==1.2.2
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.3
pytest-cov==6.0.0
pytest-django==4.9.0
sqlparse==0.5.1
tomli==2.1.0
typing_extensions==4.12.2
